Proper Note Off Handling
The MIDI spec has 0x80 being a note off.  The old Arduino MIDI lib
apparently had it so that note-offs were just note-ons with 0
velocity, but that is incorrect, and the new lib has 0x80 instead
of 0x90 for note offs.  So, registering a listener for the note
offs will fix the problem of note offs being dropped, and then
we're on to bigger and better problems!